requirements.txt
Pillow~=9.1.1
pynput~=1.7.6
PyQt5~=5.15.6